+++ fill-the-blanks/RELEASE_NOTES.md
@@ -72,3 +72,13 @@
 **2.6** | 05/08/2021
 
 * Fix formatting for multiple fields (thanks @matt2112)
+
+**2.7** | 10/09/2023 
+
+* Replace integration with Anki for compatibility with other add-ons (thanks @Abdo)
+* Possibly already make integration with other add-ons work (issues #109, #96, #115)
+* Fix feedback not showing in recent Anki (issues #109, #122, #124)
+* Update anki reference to note_type (issue #117)
+* Add warning in Template editor (issue #120)
+* Fix issue when there are many cloze or text is very long (issues #100, #116)
+* Change a bit input clean up to avoid user error (issue #111)